Extract properties into separate file
